Karel-SHC/requirements.txt
torch
Pillow
matplotlib
numpy
scipy
pandas
tqdm
tensorboardX==1.9
gym==0.15.4
requests
wandb
ply
imageio
h5py
colorlog
progressbar
torchvision
appdirs==1.4.4
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==1.2.2
colorlog==6.8.2
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
fonttools==4.49.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.42
gym==0.15.4
h5py==3.10.0
idna==3.6
imageio==2.34.0
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.1
pillow==10.2.0
ply==3.11
progressbar==2.5
protobuf==4.25.3
psutil==5.9.8
pyglet==1.3.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scipy==1.12.0
sentry-sdk==1.40.6
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
tensorboardX==1.9
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.3
appdirs==1.4.4
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==1.2.2
colorlog==6.8.2
contourpy==1.2.0
cycler==0.12.1
Django==5.0.2
docker-pycreds==0.4.0
fonttools==4.49.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.42
gym==0.15.4
h5py==3.10.0
idna==3.6
imageio==2.34.0
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.1
pillow==10.2.0
ply==3.11
progressbar==2.5
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
pyglet==1.3.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scipy==1.12.0
sentry-sdk==1.40.6
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
tensorboardX==1.9
torchvision
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.3
